Fix eslint on changes from PR #20
It applies the new rules to the old commits from the branch `legal`.
